requirements.txt
# requirements.txt is for production purpose

bentoml==1.0.25 # Packages machine learning models for deployment.
torchvision==0.14.1 # Provides tools for computer vision tasks using PyTorch.
joblib==1.2.0 # Saves and loads Python objects efficiently.
tqdm==4.64.1 # Creates progress bars for iterables.
pip-chill==1.0.1 # Installs Python packages from GitHub repositories.
wincertstore==0.2 # Interacts with the certificate store on Windows systems.
boto3
mypy-boto3-s3
streamlit 

#-e .
